Add test and fix resolving included resources
Included resources were not resolved correctly. The issue was that
https://github.com/joshdholtz/jsonapi-ios/blob/12b08a469a5ed3e27cd8b5872
11906ec3a495c75/Classes/JSONAPIResourceParser.m#L276-L276 was using the
`Class` returned by `resourceType` as a key, however the key in the
dictionary is in fact the `type` string returned by the descriptor.
Furthermore the value was set using `set:withDictionary:` which fails
because the included resource is already parsed by that point.

I also added a test confirming the bug as well as the fix.